6.4 Modified product price field type form integer to decimal

+++ apps/store/models.py
@@ -13,7 +13,7 @@ class Product(models.Model):
     slug            = models.SlugField(max_length=200, unique=True)
     category        = models.ForeignKey(Category, on_delete=models.CASCADE)
     description     = models.TextField(max_length=500, blank=True)
-    price           = models.IntegerField()
+    price 			= models.DecimalField(max_digits=10, decimal_places=2)
     images          = models.ImageField(upload_to='photos/products')
     stock           = models.IntegerField()
     is_available    = models.BooleanField(default=True)
